source code/requirements.txt
pywhatkit #For ascii art generator
ffmpeg-python #For audio/video
Pillow #Converting webp to png
argparse #Parsing arguments for commandline version
PySimpleGUI #GUI
pytube #Youtube downloader